pr 14 is completed

+++ controllers/cards.js
@@ -2,71 +2,56 @@ const path = require('path');
 
 const Card = require(path.join(__dirname, '../models/card'));
 
-const getAllCards = (req, res, next) => {
+const getAllCards = (req, res) => {
   Card.find({})
-    .then((cards) => {
-      res.status(200).send(cards);
-    })
-    .catch(next);
+    .then((cards) => res.status(200).send(cards))
+    .catch((err) => res.status(500).send(`Произошла ошибка : ${err.name}`));
 };
 
-const removeCardByID = (req, res, next) => {
-  Card.findOneAndDelete({ _id: req.params.id })
+const removeCardByID = (req, res) => {
+  Card.findOne({ _id: req.params.id })
     .orFail()
     .then((card) => {
-      const message = 'Карточка успешно удалена';
-      res.status(200).send({ message, card });
+      if (req.user._id === card.owner.toString()) {
+        Card.findOneAndDelete({ _id: req.params.id })
+          .then(() => res.status(200).send(`Карточка ${req.params.id} успешно удалена`));
+      } else {
+        res.status(403).send('Невозможно удалить чужую карточку');
+      }
     })
     .catch((err) => {
-      err.message = `Невозможно удалить карточку: ${err.message}`;
-      next(err);
+      const errStatus = (err.name === 'DocumentNotFoundError') ? 404 : 400;
+      res.status(errStatus).send(`Невозможно удалить карточку: ${err.name}`);
     });
 };
 
-const createCard = (req, res, next) => {
+const createCard = (req, res) => {
   const { name, link } = req.body;
   const owner = req.user._id;
   Card.create({ name, link, owner })
-    .then((card) => {
-      res.status(200).send(card);
-    })
+    .then((card) => res.status(200).send(card))
+    .catch((err) => res.status(400).send(`Ошибка создания карточки : ${err.name}`));
+};
+
+const likeCard = (req, res) => {
+  const owner = req.user._id;
+  Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: owner } }, { new: true })
+    .orFail()
+    .then((card) => res.status(200).send(card))
     .catch((err) => {
-      err.message = `Ошибка создания карточки: ${err.message}`;
-      next(err);
+      const errStatus = (err.name === 'DocumentNotFoundError') ? 404 : 400;
+      res.status(errStatus).send(`Ошибка добавления лайка: ${err.name}`);
     });
 };
 
-const likeCard = (req, res, next) => {
+const dislikeCard = (req, res) => {
   const owner = req.user._id;
-  Card.findByIdAndUpdate(
-    req.params.cardId,
-    { $addToSet: { likes: owner } },
-    { new: true },
-  )
+  Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: owner } }, { new: true })
     .orFail()
-    .then((card) => {
-      res.status(200).send(card);
-    })
+    .then((card) => res.status(200).send(card))
     .catch((err) => {
-      err.message = `Ошибка добавления лайка: ${err.message}`;
-      next(err);
-    });
-};
-
-const dislikeCard = (req, res, next) => {
-  const owner = req.user._id;
-  Card.findByIdAndUpdate(
-    req.params.cardId,
-    { $pull: { likes: owner } },
-    { new: true },
-  )
-    .orFail()
-    .then((card) => {
-      res.status(200).send(card);
-    })
-    .catch((err) => {
-      err.message = `Ошибка снятия лайка: ${err.message}`;
-      next(err);
+      const errStatus = (err.name === 'DocumentNotFoundError') ? 404 : 400;
+      res.status(errStatus).send(`Ошибка снятия лайка: ${err.name}`);
     });
 };
